added an outcome function that outputs the result of the game to the user.
added a computer limit that causes the computer to stop drawing once
their total is above 17. Will actually switch that to <=, as that is
more aligned with the rules of blackjack.

created a blackjack method towards the beginning of the class that can
allow the game to play in a certain order.