created two methods for doubling down on a hand.
'first_double' targets the main hand that a user will have and gives
them the option of a double-or-nothing draw.

'split_double' serves a similar purpose, except for the hand that a user
will have if they had previously split their cards.

modified 'hit' method to include this methods for easier readability